fix: let the chunker own length policy instead of the ollama adapter
The chunker grows its per chunk size for a long document so the number of
chunks stays bounded, and the ollama adapter still cut anything over eight
thousand characters before sending it. A long resume was therefore silently
truncated again, by a different route, inside the change whose purpose was to
stop that, while the log line beside it claimed no text is dropped.

The adapter no longer rewrites what it is given. Its own over length error is
already recognised, so a chunk that really is too large now comes back as a
failure the halving ladder handles, which is what that ladder is for. The body
builder is now a pure function so a grown chunk can be asserted to arrive whole
without a network mock.

Also gates the ollama thinking parameter against the levels it accepts, the way
the other three adapters already do, and checks the embedding dimension the api
actually returns rather than trusting the request shape, since two careful
readings of the vendor reference disagreed about it.
