+++ apps/desktop/src-tauri/src/commands/ai_provider/ollama.rs
@@ -80,6 +80,20 @@ pub(super) fn ollama_family_supports_thinking(model: &str) -> bool {
     m.contains("gpt-oss") || m.contains("deepseek-r1") || m.contains("deepseek-v3.1")
 }
 
+/// Levels `think` accepts on every thinking-family model
+/// ([`ollama_family_supports_thinking`]) — Ollama's `docs/capabilities/
+/// thinking.mdx` (fetched 2026-08-03) documents one uniform `low`/`medium`/
+/// `high` string enum, not a per-model-tier table like Gemini's
+/// `thinkingLevel`. Still gated with `.contains(&effort)` on the send path
+/// below, not just the family-membership boolean — the same protection
+/// Gemini/Anthropic/OpenAI use (`effort` is stored PER PROVIDER, not per
+/// model — `preferences-store.ts` — so a stale/unrecognized value must never
+/// ship just because the CURRENT model happens to be in the thinking
+/// family). A no-op today since every thinking-family model shares this same
+/// set, but it keeps this call site correct with zero further change the day
+/// a future model needs a narrower one.
+const OLLAMA_EFFORT_LEVELS: [&str; 3] = ["low", "medium", "high"];
+
 /// Parse a non-streaming `/api/chat` response into an [`AgentTurn`]:
 /// `message.content` is the text, each `message.tool_calls[]` maps to a
 /// [`ToolCall`] (Ollama returns `function.arguments` as an already-decoded JSON
@@ -179,7 +193,7 @@ impl AiProvider for OllamaClient {
 
     fn effort_levels(&self, model: &str) -> Vec<&'static str> {
         if ollama_family_supports_thinking(model) {
-            vec!["low", "medium", "high"]
+            OLLAMA_EFFORT_LEVELS.to_vec()
         } else {
             Vec::new()
         }
@@ -413,12 +427,36 @@ pub async fn reachable_model() -> (bool, Option<String>) {
 
 /// Embed `text` with a specific Ollama embedding model. Returns a clear error
 /// (not `None`) so callers can surface why embedding failed.
+///
+/// Deliberately sends `text` AS GIVEN — no internal length cap. This used to
+/// `chars().take(8000)` here "defensively", reasoning that the shared
+/// `embed_text` path already caps to `max_embedding_input_chars` (8000 for
+/// Ollama) so it would be a no-op. That stopped being true once
+/// `bounded_split_cap` (`embed.rs`) started deliberately GROWING a chunk past
+/// 8000 for a document needing more than `MAX_CHUNKS_PER_DOCUMENT` chunks at
+/// the nominal size — a grown ~9,375-char chunk arrived here and this cap
+/// silently dropped its last ~1,375 chars, no error, every time: the exact
+/// silent-truncation defect this whole adaptive-embedding feature exists to
+/// fix, reintroduced one layer down. `embed_with` has exactly one caller
+/// (`OllamaClient::embed`, reached only through `embed_adaptive`'s chunking),
+/// so there is no direct caller to protect — length policy belongs entirely
+/// to `embed_adaptive`'s chunk-and-halve ladder now: if Ollama's real request
+/// limit is smaller than what's sent, it rejects with a real error
+/// (`is_context_length_error` already recognizes Ollama's wording), which
+/// `embed_chunk_adaptive` retries at a smaller size — the ladder doing
+/// exactly what it was built for, instead of this cap quietly mutating the
+/// caller's input first.
+/// Build the `/api/embeddings` request body. Pure + unit-tested — the ONLY
+/// place `prompt` is set, so a regression that reintroduces a length cap
+/// here (see `embed_with`'s doc comment) is caught at the body-construction
+/// level, the same way `gemini::build_embed_body`/`openai`'s chat body
+/// builders are — no HTTP mock needed to prove `text` reaches the wire whole.
+fn build_ollama_embed_body(model: &str, text: &str) -> Value {
+    json!({ "model": model, "prompt": text, "keep_alive": crate::performance::ollama_keep_alive() })
+}
+
 pub async fn embed_with(model: &str, text: &str) -> AppResult<Vec<f64>> {
-    // Defensive char-boundary-safe cap for any direct caller (avoids panics on
-    // multi-byte input). The shared `embed_text` path already caps to the provider's
-    // `max_embedding_input_chars` (8000 for Ollama), so this is a no-op there.
-    let truncated: String = text.chars().take(8000).collect();
-    let body = json!({ "model": model, "prompt": truncated, "keep_alive": crate::performance::ollama_keep_alive() });
+    let body = build_ollama_embed_body(model, text);
     let endpoint = format!("{}/api/embeddings", host());
     let resp = super::retry::send_with_retry(|| {
         crate::net::http::shared()
@@ -825,15 +863,16 @@ fn build_chat_stream_body(req: &AiGenerateRequest) -> Value {
 
     let mut body = json!({ "model": req.model, "messages": messages, "stream": true });
     // `think` is a top-level request field (NOT nested under `options`), and only
-    // safe to send on a model that actually advertises thinking — it 400s on a
-    // non-thinking model (see `ollama_family_supports_thinking`'s doc comment).
+    // safe to send when it's one of `OLLAMA_EFFORT_LEVELS` (see its doc comment)
+    // — it 400s on a non-thinking model (see `ollama_family_supports_thinking`'s
+    // doc comment) or an unrecognized value.
     if let Some(effort) = req
         .effort
         .as_deref()
         .map(str::trim)
         .filter(|e| !e.is_empty())
     {
-        if ollama_family_supports_thinking(&req.model) {
+        if ollama_family_supports_thinking(&req.model) && OLLAMA_EFFORT_LEVELS.contains(&effort) {
             body["think"] = json!(effort);
         }
     }
@@ -974,324 +1013,5 @@ async fn complete_impl(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        build_chat_stream_body, normalize_show, ollama_family_supports_thinking,
-        ollama_supports_tools, parse_ollama_frames, parse_ollama_turn, parse_ollama_usage,
-        parse_web_search, OllamaClient, StreamPiece,
-    };
-    use crate::commands::ai_provider::{AiGenerateRequest, AiProvider, StopReason, ToolCall};
-    use crate::ipc_contracts::ai::AiGenerateRequestMessage;
-    use serde_json::json;
-
-    fn base_request() -> AiGenerateRequest {
-        AiGenerateRequest {
-            model: "llama3.1:8b".to_string(),
-            messages: vec![AiGenerateRequestMessage {
-                role: "user".to_string(),
-                content: "hi".to_string(),
-            }],
-            locale: "en".to_string(),
-            temperature: Some(0.8),
-            top_p: None,
-            frequency_penalty: None,
-            presence_penalty: None,
-            repeat_penalty: None,
-            max_tokens: None,
-            context_window: None,
-            effort: None,
-        }
-    }
-
-    #[test]
-    fn chat_stream_body_serializes_top_p_and_repeat_penalty_when_set() {
-        let mut req = base_request();
-        req.top_p = Some(0.95);
-        req.repeat_penalty = Some(1.15);
-        let body = build_chat_stream_body(&req);
-        assert_eq!(body["options"]["top_p"], json!(0.95));
-        assert_eq!(body["options"]["repeat_penalty"], json!(1.15));
-        // frequency_penalty is never remapped into Ollama's repeat_penalty field.
-        assert!(body["options"].get("frequency_penalty").is_none());
-    }
-
-    #[test]
-    fn chat_stream_body_omits_sampling_options_when_none() {
-        let body = build_chat_stream_body(&base_request());
-        assert!(body["options"].get("top_p").is_none());
-        assert!(body["options"].get("repeat_penalty").is_none());
-    }
-
-    #[test]
-    fn parse_ollama_frames_splits_thinking_and_content() {
-        let mut buf = String::from(
-            "{\"message\":{\"thinking\":\"hmm\"},\"done\":false}\n\
-             {\"message\":{\"content\":\"hello\"},\"done\":false}\n",
-        );
-        let pieces = parse_ollama_frames(&mut buf);
-        assert_eq!(
-            pieces,
-            vec![StreamPiece::thinking("hmm"), StreamPiece::text("hello")]
-        );
-        assert!(buf.is_empty());
-    }
-
-    #[test]
-    fn parse_ollama_frames_done_carries_final_content() {
-        // The `done:true` object becomes a sentinel carrying its final content.
-        let mut buf = String::from("{\"message\":{\"content\":\"end\"},\"done\":true}\n");
-        assert_eq!(
-            parse_ollama_frames(&mut buf),
-            vec![StreamPiece::done("end")]
-        );
-    }
-
-    #[test]
-    fn parse_ollama_frames_done_carries_real_token_usage() {
-        let mut buf = String::from(
-            "{\"message\":{\"content\":\"end\"},\"done\":true,\"prompt_eval_count\":123,\"eval_count\":45}\n",
-        );
-        let pieces = parse_ollama_frames(&mut buf);
-        assert_eq!(pieces.len(), 1);
-        let usage = pieces[0].usage.expect("done piece must carry usage");
-        assert_eq!(usage.input_tokens, 123);
-        assert_eq!(usage.output_tokens, 45);
-    }
-
-    #[test]
-    fn parse_usage_reads_prompt_eval_and_eval_counts() {
-        let data = json!({ "prompt_eval_count": 10, "eval_count": 20 });
-        let usage = parse_ollama_usage(&data).expect("usage present");
-        assert_eq!(usage.input_tokens, 10);
-        assert_eq!(usage.output_tokens, 20);
-    }
-
-    #[test]
-    fn parse_usage_is_none_when_absent() {
-        assert!(parse_ollama_usage(&json!({})).is_none());
-    }
-
-    #[test]
-    fn parse_ollama_frames_buffers_partial_line() {
-        // A partial trailing JSON line is left for the next chunk.
-        let mut buf = String::from("{\"message\":{\"content\":\"hi\"},\"done\":false}\n{\"mess");
-        let pieces = parse_ollama_frames(&mut buf);
-        assert_eq!(pieces, vec![StreamPiece::text("hi")]);
-        assert_eq!(buf, "{\"mess");
-    }
-
-    #[test]
-    fn parse_ollama_frames_skips_blank_and_unparseable_lines() {
-        let mut buf = String::from("\nnot-json\n");
-        assert!(parse_ollama_frames(&mut buf).is_empty());
-    }
-
-    #[test]
-    fn parse_web_search_maps_results_and_caps_limit() {
-        let body = json!({
-            "results": [
-                { "title": "Acme — Wikipedia", "url": "https://w/a", "content": "Acme makes widgets." },
-                { "title": "Acme careers", "url": "https://a/c", "content": "Series B." },
-                { "title": "extra", "url": "https://x", "content": "ignored by limit" },
-            ]
-        });
-        let out = parse_web_search(&body, 2);
-        assert_eq!(out.len(), 2);
-        assert_eq!(out[0].title, "Acme — Wikipedia");
-        assert_eq!(out[0].snippet, "Acme makes widgets.");
-        assert_eq!(out[1].url, "https://a/c");
-    }
-
-    #[test]
-    fn parse_web_search_tolerates_missing_fields_and_no_results() {
-        assert!(parse_web_search(&json!({}), 5).is_empty());
-        let out = parse_web_search(&json!({ "results": [{}] }), 5);
-        assert_eq!(out.len(), 1);
-        assert_eq!(out[0].title, "");
-        assert_eq!(out[0].snippet, "");
-    }
-
-    #[test]
-    fn normalize_extracts_context_and_details_by_architecture() {
-        // `context_length` is keyed by architecture — scan for the suffix, not a
-        // hardcoded `llama.` prefix, so qwen2/phi3/etc. all work unchanged.
-        let data = json!({
-            "model_info": { "qwen2.context_length": 32768, "qwen2.embedding_length": 3584 },
-            "details": { "parameter_size": "7.6B", "quantization_level": "Q4_K_M", "family": "qwen2" }
-        });
-        let out = normalize_show(&data);
-        assert_eq!(out["contextLength"], json!(32768));
-        assert_eq!(out["parameterSize"], json!("7.6B"));
-        assert_eq!(out["quantization"], json!("Q4_K_M"));
-        assert_eq!(out["family"], json!("qwen2"));
-    }
-
-    #[test]
-    fn normalize_omits_missing_fields() {
-        let data = json!({
-            "model_info": { "llama.context_length": 8192 },
-            "details": { "parameter_size": "8B" }
-        });
-        let out = normalize_show(&data);
-        assert_eq!(out["contextLength"], json!(8192));
-        assert_eq!(out["parameterSize"], json!("8B"));
-        // Absent fields are omitted (not null), so the TS optional schema accepts it.
-        assert!(out.get("quantization").is_none());
-        assert!(out.get("family").is_none());
-    }
-
-    #[test]
-    fn normalize_returns_null_when_nothing_usable() {
-        assert!(normalize_show(&json!({})).is_null());
-        assert!(normalize_show(&json!({ "model_info": {}, "details": {} })).is_null());
-    }
-
-    #[test]
-    fn tool_support_gate_is_conservative() {
-        for m in [
-            "llama3.1:8b",
-            "llama3.3:70b",
-            "qwen2.5:7b",
-            "mistral-nemo",
-            "command-r-plus",
-        ] {
-            assert!(ollama_supports_tools(m), "{m} should advertise tools");
-        }
-        // Unknown / non-tool families default off so the turn degrades safely.
-        for m in [
-            "llama2",
-            "phi3",
-            "gemma2",
-            "nomic-embed-text",
-            "deepseek-coder",
-        ] {
-            assert!(!ollama_supports_tools(m), "{m} must default to no tools");
-        }
-    }
-
-    #[test]
-    fn parse_turn_reads_object_arguments_and_content() {
-        // Ollama returns arguments as an already-decoded object (NOT a JSON string).
-        let data = json!({
-            "message": {
-                "role": "assistant",
-                "content": "",
-                "tool_calls": [{ "function": { "name": "match_resume", "arguments": { "resumeId": "r1", "jobId": "j1" } } }]
-            },
-            "done": true,
-            "done_reason": "stop"
-        });
-        let turn = parse_ollama_turn(&data);
-        assert_eq!(turn.stop, StopReason::ToolUse);
-        assert_eq!(
-            turn.tool_calls,
-            vec![ToolCall {
-                id: "match_resume-0".to_string(),
-                name: "match_resume".to_string(),
-                args: json!({ "resumeId": "r1", "jobId": "j1" }),
-            }]
-        );
-    }
-
-    #[test]
-    fn parse_turn_plain_answer_has_no_tool_calls() {
-        let data = json!({
-            "message": { "role": "assistant", "content": "The answer." },
-            "done": true,
-            "done_reason": "stop"
-        });
-        let turn = parse_ollama_turn(&data);
-        assert_eq!(turn.text, "The answer.");
-        assert!(turn.tool_calls.is_empty());
-        assert_eq!(turn.stop, StopReason::End);
-    }
-
-    #[test]
-    fn parse_turn_tool_calls_with_length_done_reason_maps_to_length_not_tool_use() {
-        // `done_reason: "length"` means the arguments may be truncated JSON — this
-        // must win over the tool-call signal, never `ToolUse`.
-        let data = json!({
-            "message": {
-                "role": "assistant",
-                "content": "",
-                "tool_calls": [{ "function": { "name": "match_resume", "arguments": { "resumeId": "r1" } } }]
-            },
-            "done": true,
-            "done_reason": "length"
-        });
-        let turn = parse_ollama_turn(&data);
-        assert_eq!(turn.stop, StopReason::Length);
-    }
-
-    #[test]
-    fn thinking_family_gate_matches_documented_models_only() {
-        for m in [
-            "qwen3",
-            "qwen3:30b",
-            "qwen3:8b",
-            "my-qwen3",
-            "gpt-oss:20b",
-            "gpt-oss:120b",
-            "deepseek-r1",
-            "deepseek-r1:70b",
-            "deepseek-v3.1:671b",
-        ] {
-            assert!(ollama_family_supports_thinking(m), "{m} should think");
-        }
-        // qwen3-coder(-plus) matches the "qwen3" substring but is a separate,
-        // non-thinking model — must NOT be swept in by a broad substring match.
-        for m in [
-            "qwen3-coder:480b",
-            "qwen3-coder-plus",
-            "llama3.1:8b",
-            "mistral",
-            "nomic-embed-text",
-        ] {
-            assert!(!ollama_family_supports_thinking(m), "{m} must not think");
-        }
-    }
-
-    #[test]
-    fn coder_exclusion_is_scoped_to_qwen3_not_every_model_containing_coder() {
-        // The exclusion is narrow — a hypothetical thinking-capable "coder"
-        // variant in an UNRELATED family must not be swept out by name
-        // collision alone (only the qwen3 branch excludes "coder").
-        assert!(ollama_family_supports_thinking("gpt-oss-coder"));
-        assert!(ollama_family_supports_thinking("deepseek-r1-coder"));
-    }
-
-    #[test]
-    fn effort_levels_mirror_the_thinking_gate() {
-        assert_eq!(
-            OllamaClient.effort_levels("gpt-oss:120b"),
-            vec!["low", "medium", "high"]
-        );
-        assert!(OllamaClient.effort_levels("llama3.1:8b").is_empty());
-    }
-
-    #[test]
-    fn chat_stream_body_sends_think_only_for_a_thinking_model_with_effort_set() {
-        let mut req = base_request();
-        req.model = "gpt-oss:20b".to_string();
-        req.effort = Some("high".to_string());
-        let body = build_chat_stream_body(&req);
-        assert_eq!(body["think"], json!("high"));
-    }
-
-    #[test]
-    fn chat_stream_body_omits_think_for_a_non_thinking_model_even_with_effort_set() {
-        let mut req = base_request();
-        req.model = "llama3.1:8b".to_string();
-        req.effort = Some("high".to_string());
-        let body = build_chat_stream_body(&req);
-        assert!(body.get("think").is_none());
-    }
-
-    #[test]
-    fn chat_stream_body_omits_think_when_effort_not_set() {
-        let mut req = base_request();
-        req.model = "gpt-oss:20b".to_string();
-        let body = build_chat_stream_body(&req);
-        assert!(body.get("think").is_none());
-    }
-}
+#[path = "ollama_tests.rs"]
+mod tests;
